Add 3 layer controller, services, repository, of search property

+++ src/main/java/com/javaweb/model/ListingDTO.java
@@ -1,53 +1,64 @@
 package com.javaweb.model;
 
+
 public class ListingDTO {
 
-    private Long listingId;
+
+    private Integer listingId;
     private String description;
     private String listingStatus;
-    private Long propertyId;
+    private Integer propertyId;
+
 
     // === Constructors ===
     public ListingDTO() {
     }
 
-    public ListingDTO(Long listingId, String description, String listingStatus, Long propertyId) {
+
+    public ListingDTO(Integer listingId, String description, String listingStatus, Integer propertyId) {
         this.listingId = listingId;
         this.description = description;
         this.listingStatus = listingStatus;
         this.propertyId = propertyId;
     }
 
+
     // === Getters & Setters ===
-    public Long getListingId() {
+    public Integer getListingId() {
         return listingId;
     }
 
-    public void setListingId(Long listingId) {
+
+    public void setListingId(Integer listingId) {
         this.listingId = listingId;
     }
+
 
     public String getDescription() {
         return description;
     }
 
+
     public void setDescription(String description) {
         this.description = description;
     }
-
     public String getListingStatus() {
         return listingStatus;
     }
+
 
     public void setListingStatus(String listingStatus) {
         this.listingStatus = listingStatus;
     }
 
-    public Long getPropertyId() {
+
+    public Integer getPropertyId() {
         return propertyId;
     }
 
-    public void setPropertyId(Long propertyId) {
+
+    public void setPropertyId(Integer propertyId) {
         this.propertyId = propertyId;
     }
 }
+
